nsp3/requirements.txt
absl-py==0.12.0
aiohttp==3.7.4
anyio==2.2.0
appdirs==1.4.4
argon2-cffi==20.1.0
astroid==2.5
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.6
Babel==2.9.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
biopython==1.78
black==20.8b1
bleach==3.3.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
cycler==0.10.0
data==0.4
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
fair-esm==0.3.1
flake8==3.9.0
funcsigs==1.0.2
future==0.18.2
google-auth==1.24.0
google-auth-oauthlib==0.4.1
grpcio==1.36.1
h5py==3.2.1
idna==2.10
importlib-metadata==3.7.3
iniconfig==1.1.1
install==1.3.4
ipykernel==5.3.4
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.8.0
jedi==0.18.0
Jinja2==2.11.3
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
latex==0.7.0
lazy-object-proxy==1.6.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.7.0
multidict==5.1.0
mypy-extensions==0.4.3
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
notebook==6.2.0
nsp3==0.0.1
numpy==1.20.1
oauthlib==3.0.1
olefile==0.46
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.2
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pip==21.0.1
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.17
protobuf==3.15.6
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.0
Pygments==2.8.1
PyJWT==2.0.1
pylint==2.7.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.17.3
PySocks==1.7.1
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rope==0.18.0
rsa==4.7.2
SciencePlots==1.0.7
scipy==1.6.0
Send2Trash==1.5.0
setuptools==49.6.0.post20210108
shutilwhich==1.1.0
six==1.15.0
sniffio==1.2.0
tempdir==0.7.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
torch==1.8.0
torchvision==0.9.0
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.36.2
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
